LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-dynamic-004-expected.txt


FAIL .target 1 assert_equals:
<div class="target" data-offset-x="50" data-offset-y="10" data-expected-width="30" data-expected-height="20"></div>
width expected 30 but got 784
